Fix opening/closing for awning in Overkiz integration (#68890)
