+++ homeassistant/components/overkiz/cover_entities/generic_cover.py
@@ -14,7 +14,8 @@ from homeassistant.components.cover import (
     SUPPORT_STOP_TILT,
     CoverEntity,
 )
-from homeassistant.components.overkiz.entity import OverkizEntity
+
+from ..entity import OverkizEntity
 
 ATTR_OBSTRUCTION_DETECTED = "obstruction-detected"
 
@@ -111,55 +112,13 @@ class OverkizGenericCover(OverkizEntity, CoverEntity):
         if command := self.executor.select_command(*COMMANDS_STOP_TILT):
             await self.executor.async_execute_command(command)
 
-    @property
-    def is_opening(self) -> bool | None:
-        """Return if the cover is opening or not."""
-
-        if self.assumed_state:
-            return None
-
-        # Check if cover movement execution is currently running
-        if any(
+    def is_running(self, commands: list[OverkizCommand]) -> bool:
+        """Return if the given commands are currently running."""
+        return any(
             execution.get("device_url") == self.device.device_url
-            and execution.get("command_name") in COMMANDS_OPEN + COMMANDS_OPEN_TILT
+            and execution.get("command_name") in commands
             for execution in self.coordinator.executions.values()
-        ):
-            return True
-
-        # Check if cover is moving based on current state
-        is_moving = self.device.states.get(OverkizState.CORE_MOVING)
-        current_closure = self.device.states.get(OverkizState.CORE_CLOSURE)
-        target_closure = self.device.states.get(OverkizState.CORE_TARGET_CLOSURE)
-
-        if not is_moving or not current_closure or not target_closure:
-            return None
-
-        return cast(int, current_closure.value) > cast(int, target_closure.value)
-
-    @property
-    def is_closing(self) -> bool | None:
-        """Return if the cover is closing or not."""
-
-        if self.assumed_state:
-            return None
-
-        # Check if cover movement execution is currently running
-        if any(
-            execution.get("device_url") == self.device.device_url
-            and execution.get("command_name") in COMMANDS_CLOSE + COMMANDS_CLOSE_TILT
-            for execution in self.coordinator.executions.values()
-        ):
-            return True
-
-        # Check if cover is moving based on current state
-        is_moving = self.device.states.get(OverkizState.CORE_MOVING)
-        current_closure = self.device.states.get(OverkizState.CORE_CLOSURE)
-        target_closure = self.device.states.get(OverkizState.CORE_TARGET_CLOSURE)
-
-        if not is_moving or not current_closure or not target_closure:
-            return None
-
-        return cast(int, current_closure.value) < cast(int, target_closure.value)
+        )
 
     @property
     def extra_state_attributes(self) -> Mapping[str, Any] | None:
